Fix all x64 warnings in d3d11Wrapper.cpp
These are the last of the warnings, mostly conversion from size_t to int.
Using SIZE_T for most parameters now, as that is what the APIs take.
When necessary, just cast from size_t to DWORD, because the data can never be that big.

+++ HLSLDecompiler/DecompileHLSL.h
@@ -8,7 +8,7 @@ struct ParseParameters
 {
 	const void *bytecode;
 	const char *decompiled;
-	long decompiledSize;
+	SIZE_T decompiledSize;
 
 	bool fixSvPosition;
 	bool fixLightPosition;
